+++ runtime/il2cpp.h
@@ -43,7 +43,7 @@
 
 void il2cpp_Init();
 
-void* il2cpp_New(uint32_t sz, uint32_t typeID, int32_t isNoRef);
+void* il2cpp_New(uint32_t sz, uint32_t typeID, uint8_t isNoRef);
 void il2cpp_CheckRange(int32_t lowerBound, int32_t length, int32_t index);
 
 void il2cpp_Yield();
